+++ src/Contracts/CanProduceMessages.php
@@ -6,9 +6,7 @@ use Junges\Kafka\Config\Sasl;
 use Junges\Kafka\Message\Message;
 use Junges\Kafka\Producers\MessageBatch;
 
-/**
- * @internal
- */
+/** @internal */
 interface CanProduceMessages extends InteractsWithConfigCallbacks
 {
     /**